GVT-2418: Split preview-view.tsx to multiple files
No logic changes are expected.
