+++ ui/src/tool-panel/track-number/dialog/track-number-delete-confirmation-dialog.tsx
@@ -5,11 +5,11 @@ import * as Snackbar from 'geoviite-design-lib/snackbar/snackbar';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import dialogStyles from 'geoviite-design-lib/dialog/dialog.scss';
-import { ChangesBeingReverted } from 'preview/preview-view';
 import { PublicationRequestDependencyList } from 'preview/publication-request-dependency-list';
 import { ChangeTimes } from 'common/common-slice';
 import { revertCandidates } from 'publication/publication-api';
 import { getChangeTimes } from 'common/change-time-api';
+import { ChangesBeingReverted } from 'preview/preview-view';
 
 type TrackNumberDeleteConfirmationDialogProps = {
     changesBeingReverted: ChangesBeingReverted;
